requirements.txt
altgraph==0.17.4
ffpyplayer==4.5.1
glcontext==2.5.0
glfw==2.7.0
moderngl==5.10.0
numpy==1.26.4
packaging==24.1
Panda3D==1.10.14
panda3d-gltf==1.2.0
panda3d-simplepbr==0.12.0
pefile==2023.2.7
pillow==10.3.0
pyassimp==5.2.5
pygame==2.5.2
PyGLM==2.7.1
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pymediainfo==6.1.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyperclip==1.8.2
PyWavefront==1.3.3
pywin32-ctypes==0.2.2
screeninfo==0.8.1
typing_extensions==4.11.0
ursina==7.0.0
